# lighter.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "lighter.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/lighter/checkConnection": {
      "get": {
        "operationId": "CheckConnection",
        "responses": {
          "200": {
            "description": "(streaming responses)",
            "schema": {
              "$ref": "#/definitions/LighterGRPCColorMessage"
            }
          }
        },
        "tags": [
          "Lighter"
        ]
      }
    },
    "/v1/lighter/loadServerLog": {
      "get": {
        "operationId": "LoadServerLog",
        "responses": {
          "200": {
            "description": "(streaming responses)",
            "schema": {
              "$ref": "#/definitions/LighterGRPCLogEntry"
            }
          }
        },
        "tags": [
          "Lighter"
        ]
      }
    },
    "/v1/lighter/scheduleSwitchState": {
      "post": {
        "operationId": "ScheduleSwitchState",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/LighterGRPCConfirmation"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LighterGRPCScheduledSwitch"
            }
          }
        ],
        "tags": [
          "Lighter"
        ]
      }
    },
    "/v1/lighter/setColor": {
      "post": {
        "operationId": "SetColor",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/LighterGRPCConfirmation"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LighterGRPCColorMessage"
            }
          }
        ],
        "tags": [
          "Lighter"
        ]
      }
    },
    "/v1/lighter/switchState": {
      "post": {
        "operationId": "SwitchState",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/LighterGRPCConfirmation"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LighterGRPCStateMessage"
            }
          }
        ],
        "tags": [
          "Lighter"
        ]
      }
    }
  },
  "definitions": {
    "LighterGRPCColorMessage": {
      "type": "object",
      "properties": {
        "b": {
          "type": "integer",
          "format": "int32"
        },
        "deviceID": {
          "type": "string",
          "format": "string"
        },
        "g": {
          "type": "integer",
          "format": "int32"
        },
        "onstate": {
          "type": "boolean",
          "format": "boolean"
        },
        "opacity": {
          "type": "integer",
          "format": "int32"
        },
        "password": {
          "type": "string",
          "format": "string"
        },
        "r": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "LighterGRPCConfirmation": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "LighterGRPCLogEntry": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "format": "string"
        },
        "time": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "LighterGRPCLogLevel": {
      "type": "string",
      "enum": [
        "DEBUG",
        "INFO",
        "WARN",
        "ERROR",
        "PANIC"
      ],
      "default": "DEBUG"
    },
    "LighterGRPCLogRequest": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "integer",
          "format": "int32"
        },
        "logLevel": {
          "$ref": "#/definitions/LighterGRPCLogLevel"
        },
        "password": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "LighterGRPCRequest": {
      "type": "object",
      "properties": {
        "deviceID": {
          "type": "string",
          "format": "string"
        },
        "password": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "LighterGRPCScheduledSwitch": {
      "type": "object",
      "properties": {
        "deviceID": {
          "type": "string",
          "format": "string"
        },
        "onstate": {
          "type": "boolean",
          "format": "boolean"
        },
        "password": {
          "type": "string",
          "format": "string"
        },
        "time": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "LighterGRPCStateMessage": {
      "type": "object",
      "properties": {
        "deviceID": {
          "type": "string",
          "format": "string"
        },
        "onstate": {
          "type": "boolean",
          "format": "boolean"
        },
        "password": {
          "type": "string",
          "format": "string"
        }
      }
    }
  }
}